requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
google-genai